delete the AbstractSearchActivity
