+++ app/src/main/java/net/squanchy/navigation/HomeActivity.java
@@ -4,14 +4,12 @@ import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.Nullable;
 import android.support.design.widget.BottomNavigationView;
-import android.support.v7.widget.SearchView;
 
 import net.squanchy.R;
 import net.squanchy.fonts.TypefaceStyleableActivity;
 import net.squanchy.navigation.view.HomeViewPagerAdapter;
 import net.squanchy.navigation.view.NoSwipeViewPager;
 import net.squanchy.search.OnSearchClickListener;
-import net.squanchy.search.AbstractSearchActivity;
 import net.squanchy.search.SearchActivity;
 
 public class HomeActivity extends TypefaceStyleableActivity implements OnSearchClickListener {
